+++ lib/src/bones_api_entity_db_mysql.dart
@@ -123,7 +123,7 @@ class DBMySQLAdapter extends DBSQLAdapter<DBMySqlConnectionWrapper> {
     String? database = config?['database'] ?? config?['db'] ?? defaultDatabase;
     String? username =
         config?['username'] ?? config?['user'] ?? defaultUsername;
-    String? password = config?['password'] ?? config?['pass'];
+    String? password = (config?['password'] ?? config?['pass'])?.toString();
 
     minConnections ??= config?['minConnections'] ?? 1;
     maxConnections ??= config?['maxConnections'] ?? 3;